// protos/interfaces/user/UserUpdate.ts
// Original file: protos/user.proto


export interface UserUpdate {
  '_id'?: (string);
  'name'?: (string);
  'authType'?: (string);
  'role'?: (string);
  'status'?: (boolean);
  'photo'?: (string);
  'password'?: (string);
}

export interface UserUpdate__Output {
  '_id'?: (string);
  'name'?: (string);
  'authType'?: (string);
  'role'?: (string);
  'status'?: (boolean);
  'photo'?: (string);
  'password'?: (string);
}